A snapshot of development, nonfunctional.
