+++ Example.vocab.txt
@@ -1,4 +1,9 @@
-7
-German | English
----
-Vocab | Definition
+German|Enlish|score
+-|-|-
+q1|2asdasdadb|123
+e4|s5|0
+e4|s5|0
+e4|s5|0
+e4|s5|0
+e4|s5|0
+e4|s5|0
